implement WeightedRandomEngine
- It's like RandomEngine, but with "bias"
- Implementation is based on the same principle as WeightedRoundRobin
where an array is prepopulated with the values based on their weights
then a random roll just picks an index in the array.
Could have maybe implemented using maybe mapping from ranges between 0 to 1
but that was easier.
- Main test is statistical and could be enhanced to examine other ranges of weights.
A pool with 3 objects weighted 2, 3, 5 seemed to generate consistent expected
results with anything above 100 iterations
- extracted prepareWeights() from WeightedRoundRobinEngine so it can be reused